pyproject.toml
[tool.poetry]
name = "Homework 6"
version = "1.0.0"
description = "Image Resizer"
authors = ["Andrej Tsvetkov  <cvetkoandrejj@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = "^0.17.6"
python-multipart = "^0.0.5"
redis = "^4.1.4"
python-dotenv = "^0.20.0"
rq = "^1.10.1"
Pillow = "^9.1.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"
pytest-cov = "^3.0.0"
autoflake = "^1.4"
isort = "^5.10.0"
black = "^22.1"
unify = "^0.5"
flake8 = "^4.0"
pylint = "2.12.2"
mypy = "^0.940"
pytest-deadfixtures = "^2.2.1"
types-redis = "^4.2.0"
fakeredis = "^1.7.1"
requests = "^2.27.1"

[build-system]
requires = ["poetry>=1.0.0"]
build-backend = "poetry.masonry.api"
